+++ sqlparser/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.14.1-dev
 
+- Support `FROM` clauses in `UPDATE` statements
 - Fix `rank` columns of fts5 tables being misreported as integers
 
 ## 0.14.0